update src jsx file comments, delete obsolete files
