+++ src/app/auth/confirm/route.js
@@ -1,10 +1,8 @@
 /*
-route the user gets sent to from the confirmation email from supabase
+Route that the user gets sent to after clicking the link from the Supabase confirmation email.
 */
 
 // https://supabase.com/docs/guides/auth/server-side/nextjs
-
-// TODO fix broken redirect -- prob need to add token_hash param to Supabase email template
 
 import { NextResponse } from "next/server";
